헤비머신건 연사 구현
흩뿌리기 구현할 것

+++ Metal/Metal/Engine/ScreenBuffer.cs
@@ -175,7 +175,7 @@ namespace Framework.Engine
 
         private (int x, int y) PositionAdjustment((int x, int y) position)
         {
-            return ((position.x - Camera.Position.X) * 2, -(position.y - Camera.Position.Y) + ShottingGame.k_Height - 1);
+            return (((int)(position.x - Camera.Position.X) * 2), (int)(-(position.y - Camera.Position.Y) + ShottingGame.k_Height - 1));
         }
 
         public void Present()
